Refactor metadata handling in wallet and verifier modules to improve organization and reusability
